+++ client/src/components/MainContent/MainContent.tsx
@@ -2,7 +2,7 @@ import styles from "./MainContent.module.css"
 import ContentCard from "@/components/MainContent/ContentCard/ContentCard";
 import About from "@/components/MainContent/About";
 import Skills from "@/components/MainContent/Skills";
-import Education from "@/components/MainContent/Education";
+import Education from "@/components/MainContent/Education/Education";
 import Experience from "@/components/MainContent/Experience";
 import Projects from "@/components/MainContent/Projects/Projects";
 import Connect from "@/components/MainContent/Connect";
